Split out eventful sample into method calls.
