+++ app/events.py
@@ -3,12 +3,13 @@ import re
 import requests
 
 import app.search as search
-from app.clients import yelpClient
-from app.clients import googleapikey
+from app.clients import yelpClient, googleapikey, eventfulkey
 
 CALENDAR_URL = 'https://www.googleapis.com/calendar/v3/calendars/{}/events?key={}'
 KONA_LAT_LONG = { 'lat': 19.622345, 'lng': -155.665041 }
 KONA_RADIUS = 100000
+
+EVENTFUL_URL = 'https://api.eventful.com/json/events/search'
 
 '''
   Best effort to fetch event details (including a Yelp id) from a Google calendar.
@@ -91,6 +92,32 @@ def getGcalEventObj(event):
             print('Searching by name, error: {}'.format(err))
 
 nameAddressRegex = re.compile('^([^0-9]*)([0-9]*.*)')
+
+def fetchEventsFromLocation(latlong, maxResults, radius):
+    params = { 'app_key': eventfulkey,
+               'units': 'mi',
+               'date': 'Today',
+               'sort_order': 'relevance',
+               'location': latlong,
+               'page_size': maxResults,
+               'within': radius }
+
+    r = requests.get(EVENTFUL_URL, params)
+    eventList = r.json()['events']['event']
+    return eventList
+
+def getEventfulEventObj(event):
+    location = { 'lat': event['latitude'], 'lng': event['longitude'] }
+    # print(str(event['venue_name']) + '_' + str(event['venue_address']) + '_' + str(event['title']) + '_\n' + str(event['description']) + '\n')
+    yelpId = search._guessYelpId(event['venue_name'], location['lat'], location['lng'])
+    if yelpId:
+        eventObj = { 'yelp_id': yelpId,
+                     'location': location,
+                     'event_summary': event['title'],
+                     'start_time': event['start_time'],
+                     'url': event['url']
+        }
+        return eventObj
 
 def getNameAndAddress(rawLocation):
     output = rawLocation.lower() \
